requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-Mail==0.9.0
Flask-Security==1.7.3
Flask-Assets==0.10
Flask-peewee==0.6.6
itsdangerous==0.24
psycopg2==2.5.3